==> 02-RulebasedAnalyzer.ipynb <==
# Word Tokenization

import nltk
from nltk.tokenize import word_tokenize,sent_tokenize

nltk.download('all')

word_tokenize("Hello World!!")

# Sentence Tokenization

sent_tokenize("Hello!!! I am doing this for manning project. I believe this helps me to work with project there by learning!!")

# scoring of the reviews

from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

analyser = SentimentIntensityAnalyzer()

def print_sentiment_scores(sentence):
    snt = analyser.polarity_scores(sentence)
    print("{:-<40} {}".format(sentence, str(snt)))

print_sentiment_scores("I just got a call from my boss - does he realise it's Saturday?")

print_sentiment_scores("I just got a call from my boss - does he realise it's Saturday? :(")

import pandas as pd

df = pd.read_csv('data/small_corpus.csv')

df['reviews']= df['reviews'].astype(str)
df['scores']=df['reviews'].apply(lambda review: analyser.polarity_scores(review))

df['compound']  = df['scores'].apply(lambda score_dict: score_dict['compound'])

df['comp_score'] = df['compound'].apply(lambda c: 'pos' if c >=0 else 'neg')
