==> 04-Tokenization.ipynb <==
from transformers import DistilBertTokenizer
tokenizer=DistilBertTokenizer.from_pretrained('distilbert-base-uncased')
len(tokenizer.vocab)

with open("vocabulary.txt", 'w') as f:
    
    # For each token...
    for token in tokenizer.vocab.keys():
        
        # Write it out and escape any unicode characters.            
        f.write(token + '\n')

import csv

with open('vocabulary.csv', 'w') as f:
    fieldnames = ['words']
    writer = csv.DictWriter(f,fieldnames);
    writer.writeheader()
    for token in tokenizer.vocab.keys():
        writer.writerow({'words': token})

token=tokenizer.encode("Can I have my Indian ")

token

tokenizer.convert_ids_to_tokens(token)
